configure accessLevels with authorization

+++ sources/resolvers/accessLevel.js
@@ -1,16 +1,30 @@
 /** @format */
 
 const queries = {};
+const ERRORS = require('../errors');
+const {responseResolver}=require('../helpers/apollo');
 
 const mutations = {
-	addAccessLevel: (parent, { accessLevel }, { dataSources }, info) => {
-		return dataSources.AccessLevelAPI.addAccessLevel(accessLevel);
+	addAccessLevel: (parent, { accessLevel }, { dataSources, permissions }, info) => {
+		if (permissions.find((permission) => permission == 'accessLevels.CRUD')) {
+			return dataSources.AccessLevelAPI.addAccessLevel(accessLevel);
+		} else {
+			return ERRORS.PERMISSION_DENIED;
+		}		
 	},
-	updateAccessLevel: (parent, args, { dataSources }, info) => {
-		return dataSources.AccessLevelAPI.updateAccessLevel(args);
+	updateAccessLevel: (parent, args, { dataSources, permissions }, info) => {
+		if (permissions.find((permission) => permission == 'accessLevels.CRUD')) {
+			return dataSources.AccessLevelAPI.updateAccessLevel(args);
+		} else {
+			return ERRORS.PERMISSION_DENIED;
+		}				
 	},
-	deleteAccessLevel: (parent, { id }, { dataSources }, info) => {
-		return dataSources.AccessLevelAPI.deleteAccessLevel(id);
+	deleteAccessLevel: (parent, { id }, { dataSources, permissions }, info) => {
+		if (permissions.find((permission) => permission == 'accessLevels.CRUD')) {
+			return dataSources.AccessLevelAPI.deleteAccessLevel(id);
+		} else {
+			return ERRORS.PERMISSION_DENIED;
+		}		
 	},
 };
 
@@ -30,6 +44,7 @@ const fieldResolvers = {
 			return await dataSources.ClubAPI.getClubById(parent.club);
 		},
 	},
+	AccessLevelResponse:responseResolver("AccessLevel")
 };
 
 module.exports = { queries, mutations, fieldResolvers };
